Revert's the healthchare_visits.py rename for review purposes
