+++ src/vivarium_nih_us_cvd/components/__init__.py
@@ -1,5 +1,5 @@
 from .disease import IschemicStroke, MyocardialInfarction
-from .healthcare_utilization import HealthcareUtilization
+from .healthcare_visits import HealthcareUtilization
 from .observers import (
     ContinuousRiskObserver,
     HealthcareVisitObserver,
